pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = 'domainck'
version = '1.2.0'
description = "Domain WHOIS checker"
readme = "README.md"
authors = [
  { name = 'Tim Parenti', email = 'timparenti@gmail.com' }
]
requires-python = '>=3.11'
dependencies = [
  'python-whois ~= 0.9.6', # https://github.com/richardpenman/whois
  'GitPython ~= 3.1.45' # https://github.com/gitpython-developers/GitPython
]

[project.optional-dependencies]
dev = [
  'mypy ~= 1.19' # https://github.com/python/mypy
]

[project.urls]
repository = "https://github.com/timparenti/domainck"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
files = ["src/domainck"]

[tool.setuptools]
# Use automatic discovery
